+++ src/pages/Transfer/CrossChainTransferForm/components/ChainSelect/ChainListModal.tsx
@@ -1,7 +1,7 @@
 import { Modal, ModalBody, ModalHeader, ModalProps } from '@/component-library/Modal';
 
-import { Chains } from './ChainInput';
 import { ChainList } from './ChainList';
+import { Chains } from './ChainSelect';
 
 type Props = {
   chains: Chains;
