+++ src/pages/Transfer/CrossChainTransferForm/components/ChainSelect/index.tsx
@@ -1,0 +1,2 @@
+export type { Chain, Chains, ChainSelectProps } from './ChainSelect';
+export { ChainSelect } from './ChainSelect';
