--- src/pages/Transfer/CrossChainTransferForm/components/SelectChainInput/index.tsx
@@ -1,2 +1,0 @@
-export type { Chain, ChainInputProps, Chains } from './ChainInput';
-export { ChainInput } from './ChainInput';
